Create interactive UI for rock paper scissors game
Previously, the RPS game logic was console based and relied on
using devtools and alert messages to use. The game logic has been
integrated into the UI of the web app. The users can now use the buttons
to choose their move and see the status of the game, the round, and
score all on the actual webpage.